+++ src/tasks/dto/search-filter.dto.ts
@@ -1,4 +1,4 @@
-import { TaskStatus } from "../tasks.model";
+import { TaskStatus } from "../task-status.enum";
 import { IsIn, IsNotEmpty, IsOptional } from 'class-validator';
 
 export class searchFilterDTO {
