+++ src/tasks/tasks.controller.ts
@@ -1,5 +1,4 @@
 import { Body, Controller, Delete, Get, Param, Patch, Post, Query, UsePipes, ValidationPipe } from '@nestjs/common';
-import { Task, TaskStatus } from './tasks.model';
 import { TasksService } from './tasks.service';
 import { createTaskDTO } from './dto/create-task.dto';
 import { searchFilterDTO } from './dto/search-filter.dto';
@@ -9,36 +8,36 @@ import { TaskStatusValidationPipe } from './pipe/task-status-validation.pipe';
 export class TasksController {
   constructor(private tasksService: TasksService) { }
 
-  @Get()
-  getTasks(@Query(ValidationPipe) filterDto: searchFilterDTO): Task[] {
-    if (Object.keys(filterDto)) {
-      return this.tasksService.getTasksWithFilters(filterDto);
-    } else {
-      return this.tasksService.getAllTasks();
-    }
-  }
+  // @Get()
+  // getTasks(@Query(ValidationPipe) filterDto: searchFilterDTO): Task[] {
+  //   if (Object.keys(filterDto)) {
+  //     return this.tasksService.getTasksWithFilters(filterDto);
+  //   } else {
+  //     return this.tasksService.getAllTasks();
+  //   }
+  // }
 
-  @Get('/:id')
-  getTaskById(@Param('id') id: string): Task {
-    return this.tasksService.getTaskbyId(id);
-  }
+  // @Get('/:id')
+  // getTaskById(@Param('id') id: string): Task {
+  //   return this.tasksService.getTaskbyId(id);
+  // }
 
-  @Post()
-  @UsePipes(ValidationPipe)
-  createTask(@Body() createTaskDTO: createTaskDTO): Task {
-    return this.tasksService.createTask(createTaskDTO);
-  }
+  // @Post()
+  // @UsePipes(ValidationPipe)
+  // createTask(@Body() createTaskDTO: createTaskDTO): Task {
+  //   return this.tasksService.createTask(createTaskDTO);
+  // }
 
-  @Delete('/:id')
-  deleteTaskById(@Param('id') id: string): void {
-    this.tasksService.deleteTaskbyid(id);
-  }
+  // @Delete('/:id')
+  // deleteTaskById(@Param('id') id: string): void {
+  //   this.tasksService.deleteTaskbyid(id);
+  // }
 
-  @Patch('/:id/status')
-  updateTaskStatus(
-    @Param('id') id: string,
-    @Body('status', TaskStatusValidationPipe) status: TaskStatus,
-  ): Task {
-    return this.tasksService.updateStatus(id, status);
-  }
+  // @Patch('/:id/status')
+  // updateTaskStatus(
+  //   @Param('id') id: string,
+  //   @Body('status', TaskStatusValidationPipe) status: TaskStatus,
+  // ): Task {
+  //   return this.tasksService.updateStatus(id, status);
+  // }
 }
